fix xss, authen, using session for admin

+++ controllers/handleDeleteHashtag.php
@@ -4,19 +4,22 @@ include_once "../models/hashtag_post.php";
 include_once "../config.php";
 include_once "../models/database.php";
 $conn = require "../inc/db.php";
+include_once "../utils/index.php";
 
-if (isset($_GET['id'])) {
-    $id = $_GET['id'];
-    try {
-        HashtagPost::deteleByHashtagId($conn, $id);
-        Hashtag::deleteById($conn, $id);
-        echo '<script>alert("Delete hashtag successfully!");
-                location.href="' . BASE_URL . '/admin/hashtag";
-            </script>';
-    } catch (\Throwable $e) {
-        echo '<script>alert("Delete hashtag failed, please try again!");
-                location.href="' . BASE_URL . '/admin/hashtag";
-            </script>
-            ';
+if (authenAdmin()) {
+    if (isset($_GET['id'])) {
+        $id = $_GET['id'];
+        try {
+            HashtagPost::deteleByHashtagId($conn, $id);
+            Hashtag::deleteById($conn, $id);
+            echo '<script>alert("Delete hashtag successfully!");
+                    location.href="' . BASE_URL . '/admin/hashtag";
+                </script>';
+        } catch (\Throwable $e) {
+            echo '<script>alert("Delete hashtag failed, please try again!");
+                    location.href="' . BASE_URL . '/admin/hashtag";
+                </script>
+                ';
+        }
     }
 }
